deepcode/requirements.txt
# LangChain / LangGraph core
langchain-core>=0.3.0
langchain-openai>=0.3.0
langchain-anthropic>=0.3.0
langchain-google-genai>=2.1.0
langgraph>=0.3.0

# MCP integration
langchain-mcp-adapters>=0.1.0
mcp>=1.0.0

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
aiofiles>=24.0.0

# PDF & document processing
PyPDF2>=3.0.0

# HTTP
httpx>=0.27.0
aiohttp>=3.10.0
